Cast: Change internal frame_id from 8 bits to 32 bits.
Needed since we will use the frame_id as a key to the crypto.
It will remain 8 bits over the wire.

Review URL: https://codereview.chromium.org/65013004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@235349 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/cast/audio_sender/audio_encoder.cc
@@ -109,8 +109,8 @@ class AudioEncoder::ImplBase {
   // call.
   int buffer_fill_end_;
 
-  // A rotating counter used to label EncodedAudioFrames.
-  uint8 frame_id_;
+  // A counter used to label EncodedAudioFrames.
+  uint32 frame_id_;
 
  private:
   DISALLOW_COPY_AND_ASSIGN(ImplBase);
