+++ media/cast/logging/logging_raw.cc
@@ -24,13 +24,13 @@ LoggingRaw::~LoggingRaw() {}
 
 void LoggingRaw::InsertFrameEvent(CastLoggingEvent event,
                                   uint32 rtp_timestamp,
-                                  uint8 frame_id) {
+                                  uint32 frame_id) {
   InsertBaseFrameEvent(event, frame_id, rtp_timestamp);
 }
 
 void LoggingRaw::InsertFrameEventWithSize(CastLoggingEvent event,
                                           uint32 rtp_timestamp,
-                                          uint8 frame_id,
+                                          uint32 frame_id,
                                           int size) {
   InsertBaseFrameEvent(event, frame_id, rtp_timestamp);
   // Now insert size.
@@ -41,7 +41,7 @@ void LoggingRaw::InsertFrameEventWithSize(CastLoggingEvent event,
 
 void LoggingRaw::InsertFrameEventWithDelay(CastLoggingEvent event,
                                            uint32 rtp_timestamp,
-                                           uint8 frame_id,
+                                           uint32 frame_id,
                                            base::TimeDelta delay) {
   InsertBaseFrameEvent(event, frame_id, rtp_timestamp);
   // Now insert delay.
@@ -50,7 +50,7 @@ void LoggingRaw::InsertFrameEventWithDelay(CastLoggingEvent event,
   it->second.delay_delta = delay;
 }
 void LoggingRaw::InsertBaseFrameEvent(CastLoggingEvent event,
-                                      uint8 frame_id,
+                                      uint32 frame_id,
                                       uint32 rtp_timestamp) {
   // Is this a new event?
   FrameRawMap::iterator it = frame_map_.find(event);
@@ -70,7 +70,7 @@ void LoggingRaw::InsertBaseFrameEvent(CastLoggingEvent event,
 
 void LoggingRaw::InsertPacketEvent(CastLoggingEvent event,
                                    uint32 rtp_timestamp,
-                                   uint8 frame_id,
+                                   uint32 frame_id,
                                    uint16 packet_id,
                                    uint16 max_packet_id,
                                    int size) {
